KE: refactor code slightly, use header for common types
